+++ packages/sv/src/cli/tests/snapshots/create-with-all-addons/svelte.config.js
@@ -1,12 +1,20 @@
 import { mdsvex } from 'mdsvex';
 import adapter from '@sveltejs/adapter-node';
+import { relative, sep } from 'node:path';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	kit: { adapter: adapter() },
-	vitePlugin: {
-		dynamicCompileOptions: ({ filename }) => ({ runes: !filename.includes('node_modules') })
+	compilerOptions: {
+		// defaults to rune mode for the project, execept for `node_modules`. Can be removed in svelte 6.
+		runes: ({ filename }) => {
+			const relativePath = relative(import.meta.dirname, filename);
+			const pathSegments = relativePath.toLowerCase().split(sep);
+			const isExternalLibrary = pathSegments.includes('node_modules');
+
+			return isExternalLibrary ? undefined : true;
+		}
 	},
+	kit: { adapter: adapter() },
 	preprocess: [mdsvex()],
 	extensions: ['.svelte', '.svx']
 };